wifi: Do not set any packet filtering
It seems that notifications are not getting through over WiFi when wlan
is suspended because of a mismatch in CAF's default packet filtering
configuration compared to OOS's (see wlan_hdd_packet_filter_rules.h).

Since CAF does not enable this by default in Cobalt, we shall disable
it as well.

Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>